Use session storage for user preferences
https://issues.redhat.com/browse/COST-4574
