+++ src/store/accountSettings/accountSettingsReducer.ts
@@ -4,16 +4,7 @@ import { FetchStatus } from 'store/common';
 import { resetState } from 'store/ui/uiActions';
 import type { ActionType } from 'typesafe-actions';
 import { getType } from 'typesafe-actions';
-import {
-  getAccountCostType,
-  getAccountCurrency,
-  isCostTypeAvailable,
-  isCurrencyAvailable,
-  setAccountCostType,
-  setAccountCurrency,
-  setCostType,
-  setCurrency,
-} from 'utils/localStorage';
+import { getAccountCurrency, isCurrencyAvailable, setAccountCurrency, setCurrency } from 'utils/sessionStorage';
 
 import {
   fetchAccountSettingsFailure,
@@ -65,9 +56,6 @@ export function accountSettingsReducer(state = defaultState, action: AccountSett
       };
     case getType(fetchAccountSettingsSuccess):
       if (action.payload.data) {
-        if (action.payload.data.cost_type) {
-          initCostType(action.payload.data.cost_type);
-        }
         if (action.payload.data.currency) {
           initCurrency(action.payload.data.currency);
         }
@@ -100,21 +88,6 @@ export function accountSettingsReducer(state = defaultState, action: AccountSett
     default:
       return state;
   }
-}
-
-// Initialize cost type in local storage
-function initCostType(value: string) {
-  // Reset UI's cost type selection if default cost type has changed.
-  const accountCostType = getAccountCostType();
-  if (accountCostType && accountCostType !== value) {
-    // Todo: remove account cost type after settings page has been moved
-    // That way, resetting the cost type for the UI should only affect the user who changed the default.
-    setCostType(accountCostType);
-  }
-  if (!isCostTypeAvailable()) {
-    setCostType(value);
-  }
-  setAccountCostType(value);
 }
 
 // Initialize currency in local storage
